:package: Move BaseKeeper to keeper.Keeper, move CLIKeeper to cli/CLIApp
