+++ session_keeper/__init__.py
@@ -1,6 +1,6 @@
 from . import version
-from .cli import main
-from .keeper import BaseKeeper, CLIKeeper
+from .cli import CLIApp
+from .keeper import Keeper
 from .storage import AbstractStorage, EncryptedJsonStorage
 
 __version__ = version.__version__
