--- session_keeper/keeper/__init__.py
@@ -1,4 +1,0 @@
-from .base import BaseKeeper
-from .cli import CLIKeeper
-
-# TODO: gui keeper class
